requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
brotli=1.0.9=hb7f2c08_9
brotli-bin=1.0.9=hb7f2c08_9
brotli-python=1.0.9=py39h7a8716b_9
bzip2=1.0.8=h0d85af4_4
ca-certificates=2023.7.22=h8857fd0_0
certifi=2023.7.22=pyhd8ed1ab_0
charset-normalizer=3.2.0=pyhd8ed1ab_0
contourpy=1.1.0=py39h8ee36c8_0
cycler=0.11.0=pyhd8ed1ab_0
filelock=3.12.2=pyhd8ed1ab_0
fonttools=4.42.0=py39hdc70f33_0
freetype=2.12.1=h3f81eb7_1
gmp=6.2.1=h2e338ed_0
gmpy2=2.1.2=py39h2da61ea_1
idna=3.4=pyhd8ed1ab_0
importlib-resources=6.0.1=pyhd8ed1ab_0
importlib_resources=6.0.1=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
kiwisolver=1.4.4=py39h92daf61_1
lcms2=2.15=h2dcdeff_1
lerc=4.0.0=hb486fe8_0
libblas=3.9.0=17_osx64_openblas
libbrotlicommon=1.0.9=hb7f2c08_9
libbrotlidec=1.0.9=hb7f2c08_9
libbrotlienc=1.0.9=hb7f2c08_9
libcblas=3.9.0=17_osx64_openblas
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.18=hac1461d_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=12_3_0_h97931a8_1
libgfortran5=12.3.0=hbd3c1fe_1
libjpeg-turbo=2.1.5.1=hb7f2c08_0
liblapack=3.9.0=17_osx64_openblas
libopenblas=0.3.23=openmp_h429af6e_0
libpng=1.6.39=ha978bb4_0
libprotobuf=3.21.12=hbc0c0cd_0
libsqlite=3.42.0=h58db7d2_0
libtiff=4.5.1=hf955e92_0
libuv=1.44.2=h0dc2134_1
libwebp-base=1.3.1=h0dc2134_0
libxcb=1.15=hb7f2c08_0
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=16.0.6=hff08bdf_0
markupsafe=2.1.3=py39hdc70f33_0
matplotlib=3.7.2=py39h6e9494a_0
matplotlib-base=3.7.2=py39h6e39af2_0
mpc=1.3.1=h81bd1dd_0
mpfr=4.2.0=h4f9bd69_0
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=hf0c8a7f_0
networkx=3.1=pyhd8ed1ab_0
nomkl=1.0=h5ca1d4c_0
numpy=1.25.2=py39h892e69a_0
openjpeg=2.5.0=h13ac156_2
openssl=3.1.2=h8a1eda9_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=py39h11b3245_1
pillow=10.0.0=py39he6683de_0
pip=23.2.1=pyhd8ed1ab_0
platformdirs=3.10.0=pyhd8ed1ab_0
pooch=1.7.0=pyha770c72_3
pthread-stubs=0.4=hc929b4f_1001
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.17=h07e1443_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.9=3_cp39
pytorch=2.0.0=cpu_generic_py39hfc57508_1
pytz=2023.3=pyhd8ed1ab_0
readline=8.2=h9e318b2_1
requests=2.31.0=pyhd8ed1ab_0
scipy=1.11.1=py39hded996c_0
setuptools=68.1.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h6db0672_2
sympy=1.12=pypyh9d50eac_103
tk=8.6.12=h5dbffcc_0
tornado=6.3.3=py39hdc70f33_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
unicodedata2=15.0.0=py39ha30fb19_0
urllib3=2.0.4=pyhd8ed1ab_0
wheel=0.41.1=pyhd8ed1ab_0
xorg-libxau=1.0.11=h0dc2134_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
zipp=3.16.2=pyhd8ed1ab_0
zstd=1.5.2=h829000d_7
